# openapi.yaml
openapi: 3.1.0
info:
  title: echsync
  description: echsync GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.echsync.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.echsync.org/product/tracking-server/).
  contact:
    name: echsync Support
    url: https://www.echsync.org/
    email: support@echsync.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.6'
servers:
  - url: https://www.echsync.org/api
    description: Demo Server 1
  - url: https://demo2.echsync.org/api
    description: Demo Server 2
  - url: https://demo3.echsync.org/api
    description: Demo Server 3
  - url: https://demo4.echsync.org/api
    description: Demo Server 4
  - url: https://server.echsync.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /permissions:
    post:
      summary: Link an Object to another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: No permission
          content: {}
      x-codegen-request-body-name: body
    delete:
      summary: Unlink an Object from another Object
      tags:
        - Permissions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Permission'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
      x-codegen-request-body-name: body
...
  /session:
    get:
      summary: Fetch Session information
      tags:
        - Session
      parameters:
        - name: token
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not Found
          content: {}
    post:
      summary: Create a new Session
      tags:
        - Session
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
    delete:
      summary: Close the Session
      tags:
        - Session
      responses:
        '204':
          description: No Content
          content: {}
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        readonly:
          type: boolean
        administrator:
          type: boolean
        map:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        zoom:
          type: integer
        password:
          type: string
        coordinateFormat:
          type: string
          nullable: true
        disabled:
          type: boolean
        expirationTime:
          type: string
          description: in IS0 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
          nullable: true
        deviceLimit:
          type: integer
        userLimit:
          type: integer
        deviceReadonly:
          type: boolean
        limitCommands:
          type: boolean
        fixedEmail:
          type: boolean
        poiLayer:
          type: string
          nullable: true
        attributes:
          type: object
          properties: {}
...
    Permission:
      type: object
      properties:
        userId:
          type: integer
          description: User id, can be only first parameter
        deviceId:
          type: integer
          description: >-
            Device id, can be first parameter or second only in combination with
            userId
        groupId:
          type: integer
          description: >-
            Group id, can be first parameter or second only in combination with
            userId
        geofenceId:
          type: integer
          description: Geofence id, can be second parameter only
        notificationId:
          type: integer
          description: Notification id, can be second parameter only
        calendarId:
          type: integer
          description: >-
            Calendar id, can be second parameter only and only in combination
            with userId
        attributeId:
          type: integer
          description: Computed attribute id, can be second parameter only
        driverId:
          type: integer
          description: Driver id, can be second parameter only
        managedUserId:
          type: integer
          description: >-
            User id, can be second parameter only and only in combination with
            userId
        commandId:
          type: integer
          description: Saved command id, can be second parameter only
      description: >-
        This is a permission map that contain two object indexes. It is used to
        link/unlink objects. Order is important. Example: { deviceId:8,
        geofenceId: 16 }
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer